# res/openapi.yaml
openapi: 3.0.0
info:
  title: grommunio Admin API
  version: 1.0.0
servers:
  - url: '/api/v1'
    description: Default deployment path
paths:
...
  /login:
    post:
      summary: Login user
      tags:
        - Misc
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                user:
                  type: string
                  description: Username
                pass:
                  type: string
                  description: User password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  grommunioAuthJwt:
                    type: string
                    description: API access token
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          description: Login failed
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
  /profile:
    get:
      summary: Get information about currently logged in user
      tags:
        - Misc
      security:
        - JWTCookie: []
      responses:
        '200':
          description: Profile data returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      username:
                        type: string
                      realName:
                        type: string
                  capabilities:
                    type: array
                    description: List of capabilities the current user has
                    items:
                      type: string
  /passwd:
    put:
      summary: Change current users password
      tags:
        - Misc
        - Service
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [new, old]
              properties:
                new:
                  type: string
                  description: The new password
                old:
                  type: string
                  description: The old password
                user:
                  type: string
                  description: Set password for a specific user (use inline authentication)
      responses:
        '200':
          description: Password updated
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          description: Authentication failed
        '403':
          description: Incorrect old password
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
...
  /system/domains:
    get:
      summary: Get lists of domains
      tags:
        - System Admin/Domains
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/verbosity'
        - $ref: '#/components/parameters/queryLimit'
        - $ref: '#/components/parameters/queryOffset'
        - $ref: '#/components/parameters/match'
        - $ref: '#/components/parameters/matchFields'
        - name: sort
          description: Sort by attribute. Can be given multiple times.
          in: query
          schema:
            type: string
            pattern: '^(ID|domainname|orgID|maxUser|title|address|adminName|tel|domainStatus)(,(a|de)sc)?$'
        - $ref: '#/components/parameters/filterID'
        - name: orgID
          description: Filter by one or more organization IDs
          in: query
          explode: false
          schema:
            $ref: '#/components/schemas/intList'
        - name: domainname
          description: Filter by domainname
          in: query
          schema:
            type: string
        - name: maxUser
          description: Filter by number of maximum users
          in: query
          explode: false
          schema:
            $ref: '#/components/schemas/intRange'
        - name: title
          description: Filter by domain title
          in: query
          schema:
            type: string
        - name: address
          description: Filter by domain address property
          in: query
          schema:
            type: string
        - name: adminName
          description: Filter by admin name
          in: query
          schema:
            type: string
        - name: tel
          description: Filter by telephone number
          in: query
          schema:
            type: string
        - name: domainStatus
          description: Filter by one or more domain status values
          in: query
          explode: false
          schema:
            $ref: '#/components/schemas/intList'
      responses:
        '200':
          description: Data returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/domain'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
    post:
      summary: Create new domain
      tags:
        - System Admin/Domains
      security:
        - JWTCookie: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/domain'
      parameters:
        - name: createRole
          description: Whether to automatically create a domain admin role for the new domain
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '201':
          description: Domain created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/domain'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
  /system/domains/{domainID}:
    get:
      summary: Get detailed info about domain
      tags:
        - System Admin/Domains
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/domainID'
        - $ref: '#/components/parameters/verbosity'
      responses:
        '200':
          description: Domain returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/domain'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
    patch:
      summary: Update domain
      tags:
        - System Admin/Domains
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/domainID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/domain'
      responses:
        '200':
          description: Domain updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/domain'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
    delete:
      summary: Delete domain
      description: Removes domain (purge=true) or marks it as deleted (purge=false)
      tags:
        - System Admin/Domains
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/domainID'
        - name: purge
          in: query
          required: false
          description: Permanently delete the domain
          schema:
            type: boolean
            default: false
        - name: deleteFiles
          in: query
          required: false
          description: Delete files from disk when purging
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Domain deleted
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
...
  /chkFormat:
    get:
      summary: Check format of input
      tags:
        - Misc
      parameters:
        - name: domain
          description: Check format of domain name
          in: query
          required: false
          schema:
            type: string
        - name: email
          description: Check format of e-mail address
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Validaty information returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  domain:
                    type: string
                    nullable: true
                    description: Error message or `null` if valid
                  email:
                    type: string
                    nullable: true
                    description: Error message or `null` if valid
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /system/orgs:
    get:
      summary: Get a list of organizations
      tags:
        - System Admin/Organizations
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/verbosity'
        - $ref: '#/components/parameters/queryLimit'
        - $ref: '#/components/parameters/queryOffset'
        - $ref: '#/components/parameters/match'
        - $ref: '#/components/parameters/matchFields'
        - name: sort
          description: Sort by attribute. Can be given multiple times.
          in: query
          schema:
            type: string
            pattern: '^(ID|name|description)(,(a|de)sc)?$'
        - name: name
          description: Filter by name
          in: query
          schema:
            type: string
        - name: description
          description: Filter by description
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of organizations returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/org'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
    post:
      summary: Create new organization
      tags:
        - System Admin/Organizations
      security:
        - JWTCookie: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/org'
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/org'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
  /system/orgs/{ID}:
    get:
      summary: Get detailed information about an organization
      tags:
        - System Admin/Organizations
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/ID'
        - $ref: '#/components/parameters/verbosity'
      responses:
        '200':
          description: LDAP configuration returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/org'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
    patch:
      summary: Update organization
      tags:
        - System Admin/Organizations
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/org'
      responses:
        '200':
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/org'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
    delete:
      summary: Delete organization
      description: Delete the organization. Associated domains are made organizationless (orgID=0)
      tags:
        - System Admin/Organizations
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Organization deleted
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
components:
  securitySchemes:
    JWTCookie:
      type: apiKey
      in: cookie
      name: grommunioAuthJwt
  parameters:
    ID:
      name: ID
      description: ID of the object
      in: path
      required: true
      schema:
        type: integer
    domainID:
      name: domainID
      description: ID of the domain
      in: path
      required: true
      schema:
        type: integer
...
    filterID:
      name: ID
      description: Filter one or more IDs
      in: query
      required: false
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/intList'
...
    match:
      name: match
      description: Match by substring
      in: query
      schema:
        type: string
    matchFields:
      name: matchFields
      description: Comma separated list of attributes to restrict matching to.
      in: query
      schema:
        type: string
    verbosity:
      name: level
      in: query
      description: Set detail level of return value. Usually ranges from 0 to 2.
      schema:
        type: integer
    queryLimit:
      name: limit
      in: query
      description: Maximum number of results to return
      schema:
        type: integer
        default: 50
    queryOffset:
      name: offset
      in: query
      description: Index of the first element to return
      schema:
        type: integer
        default: 0
...
  schemas:
    ID:
      description: Unique ID of the object
      type: integer
      readOnly: true
    intList:
      type: array
      items:
        type: integer
    intRange:
      type: array
      maxItems: 2
      items:
        type: integer
    date:
      type: string
      description: Date string
      pattern: '^\d{4}-\d{2}-\d{2}$'
...
    domain:
      type: object
      additionalProperties: false
      properties:
        ID:
          $ref: '#/components/schemas/ID'
        orgID:
          type: integer
          default: 0
        domainname:
          type: string
          description: Name of the domain, automatically converted to IDNA format
          maxLength: 64
        displayname:
          type: string
          readOnly: true
          description: Decoded unicode domain name
        homedir:
          type: string
          maxLength: 128
          readOnly: True
          default: ""
        maxUser:
          type: integer
        title:
          type: string
          maxLength: 128
          default: ""
        address:
          type: string
          maxLength: 128
          default: ""
        adminName:
          type: string
          maxLength: 32
          default: ""
        tel:
          type: string
          maxLength: 64
          default: ""
        endDay:
          $ref: '#/components/schemas/date'
        domainStatus:
          type: integer
          description: Domain status (0=Normal, 1=Suspended, 2=Out Of Date, 3=Deleted)
          default: 0
        activeUsers:
          type: integer
          readOnly: true
        inactiveUsers:
          type: integer
          readOnly: true
        syncPolicy:
          $ref: '#/components/schemas/syncPolicy'
    org:
      type: object
      properties:
        ID:
          $ref: '#/components/schemas/ID'
        name:
          type: string
        description:
          type: string
          nullable: true
...
  responses:
    ServerError:
      description: An error occured while processing the request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: String representation of the exception
    InvalidRequest:
      description: Validation of input parameters failed
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              errors:
                type: array
                description: List of errors encountered during validation
                items:
                  type: string
    DatabaseError:
      description: The database query failed
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Description of the error
...
    NotFound:
      description: The requested resource could not be found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
...